+++ bin/utils/constant_variables.py
@@ -20,10 +20,39 @@ BLAST_COLUMNS = [
     "bitscore",
 ]
 
+READ_DECLARATION = {
+    'seqs': {
+        'namespace_patterns': 'fastqc',
+        'suffix': 'R1,R2'
+    },
+    'reads after filtering': {
+        'namespace_patterns': 'fastp',
+        'suffix': 'R1+R2'
+    },
+    'unique reads': {
+        'namespace_patterns': 'humid',
+        'suffix': 'R1,R2'
+    },
+    'reads assigned': {
+        'namespace_patterns': 'kaiju',
+        'suffix': 'R1+R2'
+    }
+    #TODO:
+    # - prinseq ++
+    # - bbduk
+
+}
+
+
 CONSTRAINT_GENERAL_STATS_COLUMNS = [
-    "reads_mapped",
-    "reads_mapped_percent",
-    "reads_unmapped",
+    "(samtools Post-dedup) reads mapped %",
+    "(samtools Post-dedup) reads mapped",
+    "(samtools Post-dedup) reads unmapped %",
+    "(samtools Post-dedup) reads unmapped",
+    "(samtools Raw) reads mapped %",
+    "(samtools Raw) reads mapped",
+    "(samtools Raw) reads unmapped %",
+    "(samtools Raw) reads unmapped",
     "number_of_SNPs",
     "number_of_indels",
     "CLUSTER: mosdepth.mean_coverage",
@@ -40,17 +69,21 @@ CONSTRAINT_GENERAL_STATS_COLUMNS = [
     "(mash-screen) query-ID",
     "(mash-screen) shared-hashes",
     "(failed_mapped) mapped reads",
+    "(umitools) removed reads",
+    "(umitools) deduplicated reads"
 ]
 
 COLUMN_MAPPING = {"(blast) qlen": "consensus length", "(annotation) qlen": "consensus length"}
 
 FILES_OF_INTEREST = {
-    "samtools": "multiqc_samtools_stats",
+    "failed_mapped": "",
+    'samtools-1="samtools Raw"': "",
     "umitools": "multiqc_umitools_dedup",
-    "multiqc_general_stats": "",
     "picard": "mutliqc_picard_dups",
+    'samtools="samtools Post-dedup"': "multiqc_samtools_stats",
     "ivar_variants": "",
     "bcftools": "multiqc_bcftools_stats",
+    "multiqc_general_stats": "",
 }
 
 CLUSTER_HEADERS = {