+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Initial release of Joon-Klaps/viralgenie, created with the [nf-core](https://nf-
 - Add column in custom mpileup - Shannon entropy ([#156](https://github.com/Joon-Klaps/viralgenie/pull/156))
 - Constrain -> Constraint & further python script debugging ([#161](https://github.com/Joon-Klaps/viralgenie/pull/161))
 - include empty samples in multiqc sample overview ([#162](https://github.com/Joon-Klaps/viralgenie/pull/162))
+- Include samtools stats pre dedup & post dedup in overview tables ([#163](https://github.com/Joon-Klaps/viralgenie/pull/163))
 
 ### `Fixed`